day12 totally busted

+++ aoc24/day11/day11-2.cpp
@@ -52,7 +52,7 @@ struct Data
 	void blink()
 	{
 		
-		map<uint64_t, int> newcounts;
+		map<uint64_t, uint64_t> newcounts;
 		
 		for (auto& [k, v] : counts)
 		{
